refactor(api): Fix all the import changes from containing old api
