+++ api/opentrons/api/__init__.py
@@ -3,8 +3,8 @@ from .routers import MainRouter
 from .calibration import CalibrationManager
 
 __all__ = [
-    Session,
-    SessionManager,
-    MainRouter,
-    CalibrationManager
+    'Session',
+    'SessionManager',
+    'MainRouter',
+    'CalibrationManager'
 ]
