+++ api/opentrons/commands/commands.py
@@ -2,7 +2,8 @@ from . import types
 from ..broker import broker
 import functools
 import inspect
-from opentrons.containers import Well, Container, Slot, location_to_list
+from opentrons.legacy_api.containers\
+    import Well, Container, Slot, location_to_list
 
 
 def stringify_location(location):
