+++ api/opentrons/data_storage/serializers.py
@@ -1,5 +1,5 @@
 # pylama:ignore=E252
-from opentrons.containers.placeable import Well, Container
+from opentrons.legacy_api.containers.placeable import Well, Container
 from opentrons.util.vector import Vector
 from numbers import Number
 from typing import Tuple
