+++ api/opentrons/server/endpoints/settings.py
@@ -3,7 +3,7 @@ import shutil
 import os
 from aiohttp import web
 from opentrons.config import advanced_settings as advs
-from opentrons.robot import robot_configs as rc
+from opentrons.legacy_api.robot import robot_configs as rc
 from opentrons.data_storage import database as db
 
 log = logging.getLogger(__name__)
