+++ api/tests/opentrons/containers/test_placeable_unittest.py
@@ -1,7 +1,7 @@
 import unittest
 import math
 
-from opentrons.containers.placeable import (
+from opentrons.legacy_api.containers.placeable import (
     Container,
     Well,
     Deck,
