+++ api/tests/opentrons/integration/test_protocol.py
@@ -1,9 +1,9 @@
 import unittest
 
-from opentrons import Robot
-from opentrons.containers import load as containers_load
-from opentrons.containers.placeable import Container, Deck
-from opentrons.instruments import pipette
+from opentrons.legacy_api.robot import Robot
+from opentrons.legacy_api.containers import load as containers_load
+from opentrons.legacy_api.containers.placeable import Container, Deck
+from opentrons.legacy_api.instruments import pipette
 # TODO: Modify all calls to get a Well to use the `wells` method
