+++ api/tests/opentrons/robot/test_robots_config.py
@@ -1,5 +1,5 @@
 def test_old_probe_height(short_trash_flag):
-    from opentrons.robot import robot_configs
+    from opentrons.legacy_api.robot import robot_configs
 
     cfg = robot_configs.load()
 
@@ -8,7 +8,7 @@ def test_old_probe_height(short_trash_flag):
 
 
 def test_default_probe_height():
-    from opentrons.robot import robot_configs
+    from opentrons.legacy_api.robot import robot_configs
 
     cfg = robot_configs.load()
     assert cfg.probe_center[2] == 77.0
@@ -17,7 +17,7 @@ def test_default_probe_height():
 
 def test_load_corrupt_json():
     import os
-    from opentrons.robot import robot_configs
+    from opentrons.legacy_api.robot import robot_configs
     filename = os.path.join(os.path.dirname(__file__), 'bad_config.json')
     with open(filename, 'w') as file:
         file.write('')  # empty config file
@@ -27,7 +27,7 @@ def test_load_corrupt_json():
 
 
 def test_build_config():
-    from opentrons.robot import robot_configs
+    from opentrons.legacy_api.robot import robot_configs
 
     deck_cal = [
         [1.23, 1.23, 1.23,  1.23],
