+++ api/tests/opentrons/server/calibration_integration_test.py
@@ -3,7 +3,7 @@ from opentrons import robot
 from opentrons import deck_calibration as dc
 from opentrons.deck_calibration import endpoints
 from opentrons.trackers.pose_tracker import absolute
-from opentrons.instruments.pipette_config import Y_OFFSET_MULTI
+from opentrons.legacy_api.instruments.pipette_config import Y_OFFSET_MULTI
 
 
 # Note that several values in this file have target/expected values that do not
