+++ api/tests/opentrons/server/test_control_endpoints.py
@@ -4,7 +4,7 @@ from opentrons import robot, modules
 from opentrons.server import init
 from opentrons.drivers.smoothie_drivers.driver_3_0 import SmoothieDriver_3_0_0
 from opentrons import instruments
-from opentrons.instruments import pipette_config
+from opentrons.legacy_api.instruments import pipette_config
 
 
 async def test_get_pipettes_uncommissioned(
